Allow use of local fontello archive instead of downloading
This is controlled via the `GIRDER_LOCAL_FONTELLO_ARCHIVE` variable.
